Add neutron stable cores +2 powers in stable neutron-lib branches
Core reviewers for neutron-lib stable branches were in
neutron-lib-stable-maint group. But this group i now empty and based on
current state of neutron and neutron-lib projects, I think that we can
replace that neutron-lib-stable-maint group with neutron-stable-maint
group so neutron stable cores will have same permissions in the
neutron-lib's stable branches as well.

Change-Id: Ied0b9820cdf86505380f2392a53ffeb130d0421f