Add skyline project
99cloud has been proposing the Skyline project, a modern dashboard for
OpenStack. This has been discussed during the Xena PTG with the Horizon
team and it was decided to set it up as a parallel project for now, with
close connections to the Horizon team.

99cloud will clean up its current proof-of-concept code, post it under
two repositories (skyline-apiserver and skyline-console), start openly
collaborating around it, and in the future propose it to the TC for
future inclusion in OpenStack. In the mean time, those repositories
will be set up under the "skyline" namespace.

For more details on Skyline, see:
https://www.youtube.com/watch?v=Ro8tROYKDlE

Change-Id: Ia92fcb251f4c37510d343b1a963a6e4f8efb06a4