Drop logstash filter jobs
These jobs use complicated old school puppetry. If we're going to
maintain these logstash resources going forward we need to rewrite the
config management and the testing of the filters. We drop the testing
here as it runs aggressively in some places even though this code has
been largely static since May 2017.

Change-Id: Iedd55d1e225080cf79cbfe73182f2aab33e33db9