Clean up accessbot channels without services
In preparation to switch to setting additional channel permissions
on OFTC, drop any from our config where we don't currently provide
bot services on Freenode. In basically all cases these are retired
or otherwise abandoned channels which we continued to maintain
registrations on to prevent confusion from someone re-registering
one after deletion. We're not maintaining registrations on OFTC for
channels we know nobody is going to use, so clean them up prior to a
mass migration.

Change-Id: I6f277148641e88a974187ed48e0c04a9e1258851
Depends-On: https://review.opendev.org/792302