New Project Request: airship/image-builder
This is a request to create the airship/image-builder project in opendev.

The Airship Image Builder currently lives in the airship/images project,
and this new repository will allow us to split it out into its own
standalone project with its own lifecycle.

The new project uses the existing airship/images ACL.

Change-Id: Ie88b07af6c2be077ff0f6d8271fb6b10a364f61a