Enable publish-openstack-python-tarball job
https://review.opendev.org/c/openstack/project-config/+/791745
adds the above job, It needs to be enable for
tripleo-operator-ansible so that it will run in release
pipeline.

We defined the release job[1] with in
tripleo-operator-ansible project but it is not working
as it requires secrets which exists with in project-config repo.

For fixing and enabling it, we are adding the job here so that
it will gets triggered in the release pipeline.

[1] https://review.opendev.org/c/openstack/tripleo-operator-ansible/+/791645
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I17504632998a4ec0b4db120783852f7267c4b4c6