Use publish-to-pypi-stable-only template for deprecated repo
Deprecated deliverables only do the release on stable branches
not on master. But existing publish-to-pypi template does not
have the branch matcher to exclude the master gate from running
the release jobs.

For deprecated repo, we can use new template publish-to-pypi-stable-only
which exclude the master branch to run the release jobs. This will help
us to update the master conent which is README.rst
for example: https://review.opendev.org/c/openstack/neutron-fwaas/+/800140

This also update the deprecdated repo description to add 'DEPRECATED'.

And adding noop job for master branch so that we can merge the README.rst
update for OFTC ref.

Depends-On: https://review.opendev.org/c/openstack/openstack-zuul-jobs/+/800554
Change-Id: Id6541db65e8b7957941c72d6d29474e07604adf0