Reintroduce x/tap-as-a-service shared ACL
When x/tap-as-a-service was renamed to openstack/tap-as-a-service in
Iaa6fc7b0a9dca20ea77637a3890c603837ac7014 its ACL moved into the
openstack namespace along with it. However, because these move
changes are currently untestable, we did not spot that the ACL was
also shared by the x/tap-as-a-service-dashboard and
x/tap-as-a-service-tempest-plugin projects which were not included
in the rename work.

For now, add back the old x/tap-as-a-service Gerrit ACL as it
appeared prior to the move, so that our consistency checks will stop
failing unrelated changes. Separately, we can discuss the fate of
these other two projects with their stakeholders, and see what can
be done to clean this up a bit.

Change-Id: Iee2e37f47fee9ee1fb0016f87069d7cf90a30f5d