Update Neutron's Grafana dashboard
We made many changes in Neutron check and gate queue jobs. This patch
reflects them in the Grafana dashboard.
It also reorganize dashboard a bit by e.g. changing order of the
periodic jobs to be the same as defined in the Zuul's config file in
Neutron and consolidates some jobs like docs and pep8 into one graph.

Change-Id: I7f34d397ce6f391222e46d6c3324d25d4f5e3ed7