nodepool: drop https for ubuntu
LE recently expired it's root certificate.  Unfortunately, the base
Ubuntu distros (up to focal) use a 2019 ca-certificates package that
hasn't accounted for this.  Even more unfortunately, deboostrap can't
use the updates repo to get a more recent certificate package when
building the initial chroot.

The end result is that the chroot can't contact the mirrors via https
to actually pull the updated certificates package from them, breaking
the build.

To work around this use http:// for the mirror location.

Change-Id: Ifdb177d2ed75eab9e2acb984a9be785fba517d3e