dstat graph: update to version with fixes
This updates the on-disk cache version to the same as the default
version in zuul-jobs from I0d5e1b567c364a9e6c7aa0b95de17abffaef0434

Note we have a pull request open on this original project to
incorporate these changes, but the project has been dormat for a long
time.  If we don't have a response in the medium-term future, we can
bring this project into opendev.org git and maintain our enhancments
there.

Change-Id: I26b1a7d0dde12be7c493d44d754b9ea8f72e6e66