Add dstat_graph project
This is an import of the upstream project which has not had a commit
in 4 years.  I made a pull request that has gone uncommented for
several weeks [1].  Ergo, I think we can consider this project
inactive upstream.

With the pull request updates, it is in somewhat of a steady state.
It uses a outdated version of d3.js but otherwise is maintainable.  So
far, we've not found anything else that gives a simple stand-alone
overview of background dstat runs and can be stuffed into a Zuul
artifact quite as easily.

Import the project and we can incorporate our changes from there.

[1] https://github.com/Dabz/dstat_graph/pull/6

Change-Id: I9ee4ab5e78571c174601a408d9d14209d9c0b625