Add gerrit rename-project plugin to Zuul
We want to start experimenting with the use of this plugin instead of
our manual rename process. To do that we need to build Gerrit images
with the plugin and the first step to doing that is to have Zuul manage
the plugin repo for us.

If you are curious details about the plugin can be found at:
  https://gerrit.googlesource.com/plugins/rename-project/+/refs/heads/stable-3.3/src/main/resources/Documentation/about.md

Change-Id: I8ab8d8b70f93d8c64f54f6fc45b369375be0fdbc