Set noop jobs on opendev repos for retirement
These repos are being retired. We need to set noop jobs on the repos so
that we can land the update to the repo that marks the repos as retired.

The depends on here double checks that these repos are not needed via
puppet somehow and should be landed before we start these retirements.

Depends-On: https://review.opendev.org/c/opendev/system-config/+/829119
Change-Id: I82324a9ac6191fc9dca26adf7f8962fe6ba0b864