infra-package-needs: don't require coreutils for Rocky Linux 8
coreutils comes in two variants now, 'coreutils-single' which is a
busybox-like single binary called through symlinks and the regular
coreutils.  Both satisfy the dependency for coreutils for any other
packages, but if you explicitly ask to install coreutils over
coretuils-single you get an error.

Since coreutils-single is already in the base-image, just skip
installing it on Rocky 8.

Change-Id: I89f8cb49b0cd373e454dd37439bf6efd971233e5