Add more stats to zuul performance metrics dashboard
These stats have recently been added to Zuul, so display them.

The timing and gauage stats which represent events may be better
visualized as lines with points -- lines so we can trace the changes
between events better when there are a lot of them, and points so
we can see the events themselves.  Existing graphs which could benefit
from this are updated.

Change-Id: If3e15757ebc34ca721e102018e1e6b74e344b67f