Remove geard graphing from zuul-status dashboard
The zuul job queue is no longer stored in geard so we can safely remove
this graph which has been empty for a long time. Additionally we remove
the geard queue graph for logstash as we are beginning the process of
shutting down those services.

We update the span size of the remaining graphs to make them render a
bit more nicely without the old graphs in place.

Change-Id: I1a690cc90279547b9766c6043db6dbbe3e66deb9