Add opendev/infra-openafs-deb
This is a project to hold things related to building the OpenAFS
packages used by OpenDev infra.  These packages are pushed to a PPA
which is consumed by our production machines.

We keep a separate PPA for these packages as we have had occasions in
the past where we either needed more recent versions of OpenAFS for
various platform support, or had to apply patches with some urgency to
keep things going.  Being able to quickly produce new packages has
been very helpful.

Previously these have been created by hand, where somebody uploads the
debian package bits to the PPA manually and a .deb is produced.  Of
course the source of those packages is available, but it's not very
git-ops-y or collaborative.

This started in I07780595331da29af9f72185cbb3499210e89c19 where
fungi's concerns that I argued against were completely correct.
Trying to build across xenial -> jammy with one debian/* directory is
not going to work, and checking in multiple trees is very unwieldy in
that repo.

This proposes a separate repo where we can keep branches for each
distro's build.  We can then use the existing infrastructure we've
built to build the source packages and push to the PPA for building.

I've chosen the name with "infra-" in front to make it clear these
aren't supposed to be general purpose packages in any way, but are for
opendev infra use.  I've put this in the openstack tenant because the
build bits are still in openstack/openstack-zuul-jobs for historical
reasons -- this is where we do similar builds for RPM (which uses AFS
publishing jobs, and has openstack embedded in paths, etc.) and it
makes sense to keep them together.  We can consider moving everything
under the opendev tenant in the future.

Change-Id: Ie00c42857e7a541aa1ccb8fe480756c1e7a38dc4