Add opendev/infra-vhd-util-deb repository
Similar to the recent change
Ie00c42857e7a541aa1ccb8fe480756c1e7a38dc4, the patched version of
vhd-util in the openstack-ci PPA produces Xen (rackspace) acceptable
.vhd images from diskimage-builder.  This is another tool that isn't
really captured by our gitops model (the need for this is all covered
in [1] as to why this is a big PITA; tl;dr when porting this into qemu
things that seem like they should work don't in Rackspace, and without
being able to see the operator side of the cloud to understand what
it's actually doing it's just an exercise in frustration).

Similar to the openafs situation, the code is available from the PPA
but the history has been kept in personal repos and uploaded manually.
Having it go through gerrit is good for transparency and longer term
maintenance.

We'll have to fiddle the upload jobs, etc. so it works for both
openafs and this; so this is just the first step to get things
started.

[1] https://docs.openstack.org/diskimage-builder/latest/developer/vhd_creation.html

Change-Id: I92b9c65ade7fba2851fd798bad4f658da400dec4