nodepool: set spec_store_bypass_disable=prctl for Jammy images
We have noticed Jammy jobs failing with DISK_FULL returns from Zuul.
Upon further investigation, they have hundreds of megabytes of syslog
filled with traces.  As noted inline we've debugged this down to SSBD
interactions with some cloud providers.

Set this to the later-kernel upstream default of "prctl" to avoid this
issue.

Change-Id: I09a495a9b50c55b65221be56337f52388a5d9d48