Fix propose-updates job for requirements (3rd attempt)
The previous attempts at fixing the job were not successful yet, it
seems the variable in the pre playbook isn't overridden by a job
variable. Since the other uses of this job are running with ubuntu-focal
anyway, set the list of available python interpreters in the playbook
directly and remove the modifications from the requirements specific job
again. Also pin the nodeset for this job to guard against changes in the
base job, which will likely require another change to the interpreter
list.

[0] I9735f34639777c928fbe06febc16dd24667b1405
[1] I52cc316ce9756c4411c18c8339d44c457cb59897

Change-Id: Iddd61e1be766f2bb6da4bf62766829247fa143ce