Remove x/neutron-classifier from Zuul tenant config
This repo uses deprecated per pipeline Zuul queue: configuration. I
pushed a change [0] to correct this, but the project seems to be idle.
The last merge occured at the end of 2019. We remove this repo from the
Zuul tenant config to avoid errors when Zuul stops supporting this
deprecated syntax.

[0] https://review.opendev.org/c/x/neutron-classifier/+/842623

Change-Id: I35a6093229951659d7b80015275aebc386b4cd95