grafana: update nodepool-dib status page
Update this to new panel types and modify layout a bit.  This uses a
variable to map over the stats, rather than the old template model.

This replaces the non-working thresholds that show the "OK"/"FAILED"
with value mappings which should be more reliable.

This also includes a heatmap at the top -- it's a bit weird because
it's mapping the return values of the builds, so can range between
0->128.  The return code value doesn't really help diagnose anything,
but it can highlight if things are failing in the same way (i.e. with
the same return code).

Change-Id: I74822881251c627d8e783438fa82e39aeca5645e