Remove x/ansible-role-shade from Zuul tenant config
This repo appeas to have been integrated with windmill which was
previously removed from Zuul. The removal of windmill caused
x/ansible-role-shade to have an error in the Zuul config. Address this
by removing the role from the Zuul tenant config. Shade doesn't exist
anymore, the role hasn't been updated in years, and it seems to have
primarily been used by windmill.

Change-Id: I5061d96775ded9c43b3eafe7b2edbbc5fb500949