Set ensure_pip_virtualenv_command in Focal proposal jobs
Previously we had this error:

  No key/value pairs provided, at least one is required for this action to succeed

Which appears to be due to the default(omit) condition when setting
ensure_pip_virtualenv_command in focal proposal jobs. The issue is
default(omit) doesn't do what you think it does when defining a
variable. It is really only useful for defining module parameters that
may be omitted.

Instead we need to call ensure-tox two different ways depending on
whether or not we want to override the defaults for
ensure_pip_virtualenv_command. One method for Bionic and another for
Focal keyed off of whether or not _venv_command is defined.

Change-Id: I0cbca64f4a31c8b4eacb5e1c50f2e9fb289ce18e