project-config-grafyaml: drop job
The dependent change adds an early validation check to the
project-config-grafana change, which does the same thing as the
grafyaml test here (and then goes on to make screenshots of the
graphs, if they validate).  Remove this to aviod confusion over what
test is doing what.  Add the grafana test to the gate.

Depends-On: https://review.opendev.org/c/openstack/openstack-zuul-jobs/+/851951
Change-Id: Ia6d6d7135c3c0517ddfe27f0389ecd32978fd120