grafana: helper script updates
Responding to comments from Ibbc2b116d0c496655a7ce6bb6971e8270ac32647

Make the script run in the grafana .yaml directory so you don't have
working-directory issues.

Put the secrets in this directory and add them to .gitignore

No need for a tty on the loading call

Additionally make it executable, and have a short pause to let the
container start.

Change-Id: Icf0a2cfb1a0f5599704d6c8c9e85345d61884cd5