Move the requirements-constraints job to the periodic-weekly pipeline
Generating these updates on a daily basis isn't realistic and they often
take several modifications to merge.  The current process withing the
requirements team is to modify the topic and then work on them.  The
next time the periodic pipeline triggers we "ignore" the new change.

This change switches to the periodic-weekly pipeline to better match the
available workflow of the requirements team

Change-Id: I478a64b21d9f19ece21ec5fb7a1327a106664529