[glance] Make Review-Priority completely sticky
Currently the value persists only if it's the max or min, and only
when there are no code changes or if it's a trivial rebase.

The behavior we'd like is that the assigned priority should persist
across the series of patch sets, even if there are code changes,
because this label indicates the relative importance of getting eyes
on the review, and is not directly involved in the code being approved.

A -1 Review-Priority ("Branch Freeze"), in addition to being sticky,
will continue to block submission of the patch.

Change-Id: Iee323ab80f9a58bf49827c23934560edfad57bf3