Exclude .cache when running flake8
For local runs of testenv:linters we cache all the external Ansible
playbooks and roles ansible-lint will need to reference in order to
validate content in this repository. We build that cache in a .cache
directory, but not everything in the cached repositories is suitable
for flake8 checking, so add that directory to its exclusion list.

Also add the same directory to .gitignore, so Git will stop
complaining about it being untracked.

Change-Id: I1f740edd75cc7f6ba0f4a73f0a409a620b437b16