add fasteners to openstack tenant
This change adds the fasteners project to the github
connection in the openstack tenant. several projects
(oslo, taskflow and tooz)  use fasteners directly.

By adding fasteners to the set of projects we mirror in zuul we can use
depends on to validate that bugfixes to fasteners will work with our
openstack projects before they are merged and released.

we do not want to import job config from fasteners as we do not gate it
so its added to the existing set of non gated repos.

Related-Bug: #1988311
Change-Id: If7b1412c321e5f0fc5b38314f5e9e18b8a7bfe39