Update OpenDev tenants' default ansible to 6
Zuul is looking at dropping ansible 5 support. By updating our default
ansible version to 6 early we can provide valuable feedback to the Zuul
project. That said we expect this to function. The Zuul Zuul tenant is
already running under ansible 6, and I've done some testing with
devstack [0] that appears to show those jobs running happily.

The openstack release is now out the door which makes this a good time
to change the zuul default.

[0] https://review.opendev.org/c/openstack/devstack/+/858436

Change-Id: I61493e74d9955a444326d72e085f511a838f88dd