Add another role for Zookeeper installation
During PTG it was decided that we need to provide coordination service
as part as OpenStack-Ansible deployment and decided to stop on Zookeeper
as most feature-complete solution. With that we looked for alternative
role that's present under opendev umbrella, but according to
our experience [1] of contribution to that repo it feels tought
to try to re-use it as it's missing some aspects that we feel essential
to have. Based on that we decided to re-write the role from scratch
and maintain it additionally to already existing one.

[1] https://review.opendev.org/q/project:+windmill/ansible-role-zookeeper+AND+owner:noonedeadpunk

Needed-By: https://review.opendev.org/c/openstack/governance/+/863161
Change-Id: I83d53d801f7f0e27769fda7e1a606d90726f98b2