Use kolla.config for kolla-ansible in gerrit
The kolla project has decided that it no longer needs to have a
dedicated kolla-ansible-core group that is different from the kolla-core
group. Membership between the two groups has already been synced, so
this change is essentially a no-op, but allows us to only maintain a
single group going forward.

[0] https://meetings.opendev.org/meetings/kolla/2022/kolla.2022-11-23-14.00.log.html

Change-Id: I1dc5e5ed9f93ad40f2089891c63b530994b290f2