Fix nodepool API ops Grafana graphs
openstacksdk statsd records for API operations are not response code
specific. This change adds a glob to the statsd record path to include
all return code response metrics in our API graphs. I believe this
change may have come with the openstacksdk 0.103.0 update. We also
update paths for servers POST, server details, and flavor details as
they have changed.

Note the network info is empty which is why we don't get graphs for
them, but the paths appear correct. I think this may be beacuse we don't
need to query network info in any of our clouds currently.

While we are at it we stop updating the airship and inap cloud graphs
since those should be cleaned up and this keeps the review overhead
smaller.

Change-Id: I5a6b80118afaf3b7782a1d1c131787f208583799