Update post-review zuul pipeline definiton
This change does two things. First we move the post-review definition
after gate to keep things ordered on the dashboard more consistently.
All the empty space on the right doesn't look great.

Second we replace require-approval with approval in the gerrit trigger
def. This means we only trigger when approvals are explicitly given
rather than on every event after an approval is set. I think this gives
us a bit more control and we can relax things later if necessary. But to
start this makes more sense to me.

Change-Id: Id4f8002e34f4ae4c4fceafebd577e96f9625d4e0