arm64: remove min-ram constraints
All of the flavor names in use here should be matched exactly.

min-ram (as described at [1]) when used in combination with
flavor-name actually has the wrong semantics for us, as this chooses a
flavor that "meets providers.labels.min-ram and also contain
flavor-name".  This can cause the wrong flavor to be picked up if they
use the same prefix.

[1] https://zuul-ci.org/docs/nodepool/3.10.0/configuration.html#attr-providers.[openstack].pools.labels.flavor-name

Change-Id: Iecd0e9cfe7d7f1f28dd8d0c5ce9beb6bbdf357e8