nodepool: fix new linaro provider name in nb04
I have left the -regionone off this, making its naming inconsistent.
This adds it.

Since this cloud is in its bringup phase, I will put the builder in
emergency, clear out the images for the "linaro" provider and then
apply this by hand, so that we don't have old ZK nodes lying around.
We can then merge this to make it consistent.

Change-Id: I23328cbc53b87e1e81d26cc56f99aaad33b415c0