nodepool: size linaro max-servers to subnet
As noted inline, this has a /27 subnet allocation, so that is the real
upper limit on hosts we can simultaneously talk to.

That makes 29 usable floating ip's, minus one for the mirror node.  We
would max out the 160 CPU's with our standard 8x instances (we can fit
20 * 8) before we got to this anyway, but I think it's good to just
have a note on it.

Change-Id: I63afab1a74aa76d1eb8629b3aa4df540daffa470