Remove add_master_python3_jobs.sh
In Zed cycle, TC agreed to use the unversioned job template [1]
(implemented in [2]), thus we don't need the script to generate the
patches. This removes the script and also the lines where the script is
called from, so OpenStack Bot won't propose any change regarding the
series specific job templates.

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L360
[2] https://review.opendev.org/c/openstack/openstack-zuul-jobs/+/856903

Change-Id: I8c8aa9ae1fc395361fa6162983feac631608032f