gerrit/acl : submit-requirements for deprecated NoOp function
The submit functions are deprecated in Gerrit 3.7 and replaced with
submit-requirements.

This starts at the replacement with ACL's currently using the NoOp
function.

This implements the migrations steps encded with [1] upstream.  The
function is changed to NoBlock, and a "non-applicable"
submit-requirement of "applicableIf = is:false" and "submittableIf =
is:true" is added, with the same name as the label.

Since we are matching the upstream rules for idempotence -- a
submit-requirement with the same name as the label, we will avoid
future upgrades modifying our ACL's and getting them out of sync with
what we have in project-config.

From a user's point-of-view, this change itself should be a no-op.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/339542

Change-Id: Ic43f561174ebf30474b1b54be2bed02695cebedc