gerrit/acl : Convert Backport-Candidate to submit-requirements
Similar to Ic43f561174ebf30474b1b54be2bed02695cebedc, this converts
the existing Backport-Candidate rules to submit-requirements following
the gerrit migration rules from [1].

The function is left as NoBlock, and a no-op submit-requirement is
added with the same name.

As with the prior change; the submit-requirement with the same name as
the label will avoid us having migrations run on these ACL's and keep
our gerrit in-sync with project-config.

Also as with the prior change, this should have no affect for users.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/339542

Change-Id: I83160aeec0a450f8678ecb583fb7570ac0e71f4a