Add the main NebulOuS repos
The NebulOuS project aims to deliver a Meta Operating System and
Fog Brokerage Platform for transient cloud continuum ecosystems. [1]

This patch follows the Project Creator's Guide to create a new
project in OpenDev's collaboratory.

It also creates the project-config repo in preparation for the creation
of a new Zuul tenant: nebulous.

[1] https://www.nebulouscloud.eu/

Change-Id: I369177cd43f92f02a73011b46545210015d5112a